장바구니 컴포넌트 관련된 코드 분리
